feat(UI-1004): reorganize sidebar icons and update icon imports
