+++ src/assets/image/icons/sidebar/index.ts
@@ -1,0 +1,16 @@
+// Taken from https://www.figma.com/design/klCZL68fUd6pw0DEu0Sj7f/Autokitteh%3A-Look-%26-Feel-design-versions?node-id=3803-130733&t=7Ia4ZjBoflbDQyIb-4
+export { default as CircleQuestionIcon } from "@assets/image/icons/sidebar/CircleQuestion.svg?react";
+export { default as ConnectionsIcon } from "@assets/image/icons/sidebar/Connections.svg?react";
+export { default as DashboardIcon } from "@assets/image/icons/sidebar/Dashboard.svg?react";
+// Taken from https://www.figma.com/design/klCZL68fUd6pw0DEu0Sj7f/Autokitteh%3A-Look-%26-Feel-design-versions?node-id=3803-130721&t=7Ia4ZjBoflbDQyIb-4
+export { default as EventListIcon } from "@assets/image/icons/sidebar/EventList.svg?react";
+// Taken from https://www.figma.com/design/klCZL68fUd6pw0DEu0Sj7f/Autokitteh%3A-Look-%26-Feel-design-versions?node-id=3803-130729&t=7Ia4ZjBoflbDQyIb-4
+export { default as FileIcon } from "@assets/image/icons/sidebar/File.svg?react";
+// Taken from https://www.figma.com/design/klCZL68fUd6pw0DEu0Sj7f/Autokitteh%3A-Look-%26-Feel-design-versions?node-id=3803-130738&t=7Ia4ZjBoflbDQyIb-4
+export { default as LogoutIcon } from "@assets/image/icons/sidebar/Logout.svg?react";
+export { default as NewProject } from "@assets/image/icons/sidebar/NewProject.svg?react";
+export { default as NotificationIcon } from "@assets/image/icons/sidebar/Notification.svg?react";
+export { default as ProjectsIcon } from "@assets/image/icons/sidebar/Projects.svg?react";
+// Taken from https://www.figma.com/design/klCZL68fUd6pw0DEu0Sj7f/Autokitteh%3A-Look-%26-Feel-design-versions?node-id=3803-130736&t=7Ia4ZjBoflbDQyIb-4
+export { default as SettingsIcon } from "@assets/image/icons/sidebar/Settings.svg?react";
+export { default as StatsBlackIcon } from "@assets/image/icons/sidebar/Stats.svg?react";
