--- src/assets/image/sidebar/index.ts
@@ -1,11 +1,0 @@
-export { default as CircleQuestionIcon } from "@assets/image/sidebar/CircleQuestion.svg?react";
-export { default as ConnectionsIcon } from "@assets/image/sidebar/Connections.svg?react";
-export { default as DashboardIcon } from "@assets/image/sidebar/Dashboard.svg?react";
-export { default as FileIcon } from "@assets/image/sidebar/File.svg?react";
-export { default as FlagIcon } from "@assets/image/sidebar/Flag.svg?react";
-export { default as LogoutIcon } from "@assets/image/sidebar/Logout.svg?react";
-export { default as NewProject } from "@assets/image/sidebar/NewProject.svg?react";
-export { default as NotificationIcon } from "@assets/image/sidebar/Notification.svg?react";
-export { default as ProjectsIcon } from "@assets/image/sidebar/Projects.svg?react";
-export { default as SettingsIcon } from "@assets/image/sidebar/Settings.svg?react";
-export { default as StatsBlackIcon } from "@assets/image/sidebar/Stats.svg?react";
